@@ -1,9 +1,18 @@
-import 'package:daily_mood/components/drawer/drawer.dart';
-import 'package:daily_mood/components/moodButton/moodButton.dart';
 import 'package:flutter/material.dart';
+import 'package:daily_mood/components/drawer.dart';
+import 'package:daily_mood/components/moodButtons.dart';
 
 class Home extends StatefulWidget {
-  Home({Key key, this.title}) : super(key: key);
+  Home({Key? key, required this.title}) : super(key: key);
+
+  // This widget is the home page of your application. It is stateful, meaning
+  // that it has a State object (defined below) that contains fields that affect
+  // how it looks.
+
+  // This class is the configuration for the state. It holds the values (in this
+  // case the title) provided by the parent (in this case the App widget) and
+  // used by the build method of the State. Fields in a Widget subclass are
+  // always marked "final".
 
   final String title;
 
@@ -14,25 +23,28 @@ class Home extends StatefulWidget {
 class _HomeState extends State<Home> {
   @override
   Widget build(BuildContext context) {
+    // This method is rerun every time setState is called, for instance as done
+    // by the _incrementCounter method above.
+    //
+    // The Flutter framework has been optimized to make rerunning build methods
+    // fast, so that you can just rebuild anything that needs updating rather
+    // than having to individually change instances of widgets.
+
     return Scaffold(
       appBar: AppBar(
+        // Here we take the value from the MyHomePage object that was created by
+        // the App.build method, and use it to set our appbar title.
         title: Text(widget.title),
       ),
-      drawer: MyDrawer(),
-      body: Center(
-        child: Container(
-          margin: const EdgeInsets.only(top: 10),
-          child: Column(
-            children: [
-              Text(
-                'Comment allez vous ?',
-                style: TextStyle(fontSize: 24),
-              ),
-              MoodButton(),
-            ],
-          ),
-        ),
+      body: ListView(
+        children: <Widget>[
+          Padding(
+            padding: const EdgeInsets.only(top: 8.0),
+            child: MoobButtons(),
+          )
+        ],
       ),
+      drawer: GlobalDrawer(),
     );
   }
 }
