@@ -1,7 +1,6 @@
+import 'package:daily_mood/screens/home.dart';
+import 'package:daily_mood/screens/summary.dart';
 import 'package:flutter/material.dart';
-
-import 'routes/routes.dart';
-import 'screens/home.dart';
 
 void main() {
   runApp(MyApp());
@@ -12,27 +11,18 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Dayli Mood',
-      debugShowCheckedModeBanner: false,
+      title: 'Daily Mood',
       theme: ThemeData(
-        // This is the theme of your application.
-        //
-        // Try running your application with "flutter run". You'll see the
-        // application has a blue toolbar. Then, without quitting the app, try
-        // changing the primarySwatch below to Colors.green and then invoke
-        // "hot reload" (press "r" in the console where you ran "flutter run",
-        // or simply save your changes to "hot reload" in a Flutter IDE).
-        // Notice that the counter didn't reset back to zero; the application
-        // is not restarted.
-        primarySwatch: Colors.green,
-        // This makes the visual density adapt to the platform that you run
-        // the app on. For desktop platforms, the controls will be smaller and
-        // closer together (more dense) than on mobile platforms.
-        visualDensity: VisualDensity.adaptivePlatformDensity,
+        primarySwatch: Colors.teal,
       ),
+      debugShowCheckedModeBanner: false,
       initialRoute: '/',
-      routes: routes,
-      home: Home(title: 'Daily Mood'),
+      routes: {
+        // When navigating to the "/" route, build the FirstScreen widget.
+        '/': (context) => (Home(title: 'Daily Mood')),
+        // When navigating to the "/second" route, build the SecondScreen widget.
+        '/summary': (context) => Summary(),
+      },
     );
   }
 }
